refactor: cleanup, path alias
