+++ src/infrastructure/event-handlers/Release.ts
@@ -1,15 +1,16 @@
 import type { HandlerFunction } from "@octokit/webhooks/dist-types/types";
-import { HOME_GROUP } from "env";
 import type { Context } from "grammy";
-import type { IReleaseEvent } from "src/application/interfaces/events";
-import { markdownToHTML } from "src/utils/markdown";
 import templite from "templite";
+import { HOME_GROUP } from "~/env";
+import type { IReleaseEvent } from "~/application/interfaces/events";
+import { markdownToHTML } from "~/utils/markdown";
 
 export type ReleaseTemplate = {
   published: string;
 };
 
 export class ReleaseEventHandler implements IReleaseEvent {
+  // eslint-disable-next-line no-useless-constructor
   constructor(private readonly _templates: ReleaseTemplate) {}
 
   published(ctx: Context): HandlerFunction<"release.published", unknown> {
