+++ src/infrastructure/event-handlers/Vulnerability.ts
@@ -1,8 +1,8 @@
 import type { HandlerFunction } from "@octokit/webhooks/dist-types/types";
-import { HOME_GROUP } from "env";
 import type { Context } from "grammy";
-import type { IVulnerabilityEvent } from "src/application/interfaces/events";
 import templite from "templite";
+import { HOME_GROUP } from "~/env";
+import type { IVulnerabilityEvent } from "~/application/interfaces/events";
 
 export type VulnerabilityTemplate = {
   created: {
@@ -12,6 +12,7 @@ export type VulnerabilityTemplate = {
 };
 
 export class VulnerabilityEventHandler implements IVulnerabilityEvent {
+  // eslint-disable-next-line no-useless-constructor
   constructor(private readonly _templates: VulnerabilityTemplate) {}
 
   created(ctx: Context): HandlerFunction<"repository_vulnerability_alert.create", unknown> {
@@ -34,6 +35,7 @@ export class VulnerabilityEventHandler implements IVulnerabilityEvent {
           disable_web_page_preview: true
         });
       } catch (e) {
+        // TODO: proper logging
         console.error(e);
       }
     };
